Re-organizing backup and restore into its own screen.
- Fixes #1099
