+++ app/src/main/java/com/dessalines/thumbkey/MainActivity.kt
@@ -22,11 +22,12 @@ import com.dessalines.thumbkey.db.AppSettingsRepository
 import com.dessalines.thumbkey.db.AppSettingsViewModel
 import com.dessalines.thumbkey.db.AppSettingsViewModelFactory
 import com.dessalines.thumbkey.ui.components.common.ShowChangelog
-import com.dessalines.thumbkey.ui.components.settings.SettingsActivity
-import com.dessalines.thumbkey.ui.components.settings.about.AboutActivity
-import com.dessalines.thumbkey.ui.components.settings.behavior.BehaviorActivity
-import com.dessalines.thumbkey.ui.components.settings.lookandfeel.LookAndFeelActivity
-import com.dessalines.thumbkey.ui.components.setup.SetupActivity
+import com.dessalines.thumbkey.ui.components.settings.SettingsScreen
+import com.dessalines.thumbkey.ui.components.settings.about.AboutScreen
+import com.dessalines.thumbkey.ui.components.settings.backupandrestore.BackupAndRestoreScreen
+import com.dessalines.thumbkey.ui.components.settings.behavior.BehaviorScreen
+import com.dessalines.thumbkey.ui.components.settings.lookandfeel.LookAndFeelScreen
+import com.dessalines.thumbkey.ui.components.setup.SetupScreen
 import com.dessalines.thumbkey.ui.theme.ThumbkeyTheme
 import com.dessalines.thumbkey.utils.ANIMATION_SPEED
 import com.dessalines.thumbkey.utils.getImeNames
@@ -112,14 +113,14 @@ class MainActivity : AppCompatActivity() {
                     composable(
                         route = "setup",
                     ) {
-                        SetupActivity(
+                        SetupScreen(
                             navController = navController,
                             thumbkeyEnabled = thumbkeyEnabled,
                             thumbkeySelected = thumbkeySelected,
                         )
                     }
                     composable(route = "settings") {
-                        SettingsActivity(
+                        SettingsScreen(
                             navController = navController,
                             appSettingsViewModel = appSettingsViewModel,
                             thumbkeyEnabled = thumbkeyEnabled,
@@ -127,13 +128,13 @@ class MainActivity : AppCompatActivity() {
                         )
                     }
                     composable(route = "lookAndFeel") {
-                        LookAndFeelActivity(
+                        LookAndFeelScreen(
                             navController = navController,
                             appSettingsViewModel = appSettingsViewModel,
                         )
                     }
                     composable(route = "behavior") {
-                        BehaviorActivity(
+                        BehaviorScreen(
                             navController = navController,
                             appSettingsViewModel = appSettingsViewModel,
                         )
@@ -141,8 +142,16 @@ class MainActivity : AppCompatActivity() {
                     composable(
                         route = "about",
                     ) {
-                        AboutActivity(
+                        AboutScreen(
                             navController = navController,
+                        )
+                    }
+                    composable(
+                        route = "backupAndRestore",
+                    ) {
+                        BackupAndRestoreScreen(
+                            navController = navController,
+                            appSettingsViewModel = appSettingsViewModel,
                         )
                     }
                 }
