+++ app/src/main/res/values/strings.xml
@@ -12,14 +12,14 @@
     <string name="blue">Blue</string>
     <string name="dracula">Dracula</string>
     <string name="twilight">Twilight</string>
-    <string name="high_contrast">High Contrast</string>
-    <string name="high_contrast_colorful">High Contrast Colorful</string>
+    <string name="high_contrast">High contrast</string>
+    <string name="high_contrast_colorful">High contrast colorful</string>
     <string name="ancom">Anarcho-Communist colors</string>
     <string name="center">Center</string>
     <string name="right">Right</string>
     <string name="left">Left</string>
     <string name="about">About</string>
-    <string name="whats_new">What\'s New</string>
+    <string name="whats_new">What\'s new</string>
     <string name="version">Version %1$s</string>
     <string name="releases">Releases</string>
     <string name="support">Support</string>
@@ -41,15 +41,15 @@
     <string name="vibrate_on_tap">Vibrate on tap</string>
     <string name="vibrate_warning">Requires touch feedback to be enabled in system settings</string>
     <string name="play_sound_on_tap">Play sound on tap</string>
-    <string name="auto_capitalize">Auto-Capitalize</string>
-    <string name="key_size">Key Size: %1$s</string>
-    <string name="key_non_square">Use Non-Square Keys</string>
-    <string name="key_height">Key Height: %1$s</string>
-    <string name="key_width">Key Width: %1$s</string>
-    <string name="bottom_offset">Bottom Offset: %1$s</string>
-    <string name="min_swipe_length">Minimum Swipe Length: %1$s</string>
-    <string name="animation_speed">Animation Speed: %1$s</string>
-    <string name="animation_helper_speed">Animation Helper Speed: %1$s</string>
+    <string name="auto_capitalize">Auto-capitalize</string>
+    <string name="key_size">Key size: %1$s</string>
+    <string name="key_non_square">Use non-square keys</string>
+    <string name="key_height">Key height: %1$s</string>
+    <string name="key_width">Key width: %1$s</string>
+    <string name="bottom_offset">Bottom offset: %1$s</string>
+    <string name="min_swipe_length">Minimum swipe length: %1$s</string>
+    <string name="animation_speed">Animation speed: %1$s</string>
+    <string name="animation_helper_speed">Animation helper speed: %1$s</string>
     <string name="slide_enable">Slide gestures</string>
     <string name="slide_cursor_movement_mode">Cursor movement mode</string>
     <string name="slide_sensitivity">Slide sensitivity: %1$s</string>
@@ -65,21 +65,21 @@
     <string name="cancel">Cancel</string>
     <string name="test_out_thumbkey">Test out Thumb-Key</string>
     <string name="setup">Setup</string>
-    <string name="user_guide">User Guide</string>
+    <string name="user_guide">User guide</string>
     <string name="setup_thumbkey">Setup Thumb-Key</string>
     <string name="enable_thumbkey">Enable Thumb-Key</string>
     <string name="select_thumbkey">Select Thumb-Key</string>
-    <string name="finish_setup">Finish Setup</string>
-    <string name="copy">Copied Selected Text</string>
-    <string name="hide_letters">Hide Letters</string>
-    <string name="hide_symbols">Hide Symbols</string>
-    <string name="spacebar_multitaps">Spacebar Multi-taps</string>
+    <string name="finish_setup">Finish setup</string>
+    <string name="copy">Copied selected text</string>
+    <string name="hide_letters">Hide letters</string>
+    <string name="hide_symbols">Hide symbols</string>
+    <string name="spacebar_multitaps">Spacebar multi-taps</string>
     <string name="done">Done</string>
-    <string name="backdrop">Keyboard Backdrop</string>
-    <string name="key_padding">Key Padding: %1$s</string>
-    <string name="key_border_width">Key Border Width: %1$s</string>
-    <string name="key_radius">Key Radius: %1$s</string>
-    <string name="drag_return_enable">Drag-and-Return</string>
+    <string name="backdrop">Keyboard backdrop</string>
+    <string name="key_padding">Key padding: %1$s</string>
+    <string name="key_border_width">Key border width: %1$s</string>
+    <string name="key_radius">Key radius: %1$s</string>
+    <string name="drag_return_enable">Drag-and-return</string>
     <string name="counterclockwise_drag_action">Counterclockwise drag action</string>
     <string name="clockwise_drag_action">Clockwise drag action</string>
     <string name="send_oppsite_case">Opposite case letter</string>
@@ -87,9 +87,10 @@
     <string name="circular_drag_enable">Circular drag</string>
     <string name="ghost_keys_enable">Ghost keys</string>
     <string name="ghost_keys_description">Access hidden symbol keys without switching to numeric view</string>
-    <string name="backup_database">Backup Database</string>
-    <string name="restore_database">Restore Database</string>
+    <string name="backup_database">Backup database</string>
+    <string name="restore_database">Restore database</string>
     <string name="restore_database_warning">Warning: This will clear out your current database</string>
     <string name="database_restored">Database Restored.</string>
     <string name="database_backed_up">Database Backed up.</string>
+    <string name="backup_and_restore">Backup and restore</string>
 </resources>
